+++ OneBlink.SDK/FormsAppEnvironmentsClient.cs
@@ -19,11 +19,12 @@ namespace OneBlink.SDK
             );
         }
 
-        public async Task<FormsAppEnvironmentsSearchResult> Search(int? limit = null, int? offset = null)
+        public async Task<FormsAppEnvironmentsSearchResult> Search(int? limit = null, int? offset = null, long? workspaceId = null)
         {
             var query = HttpUtility.ParseQueryString(string.Empty);
             OneBlinkHttpClient.AddItemToQuery(query, nameof(limit), limit);
             OneBlinkHttpClient.AddItemToQuery(query, nameof(offset), offset);
+            OneBlinkHttpClient.AddItemToQuery(query, nameof(workspaceId), workspaceId);
 
             string url = "/forms-app-environments?" + query.ToString();
             return await this.oneBlinkApiClient.GetRequest<FormsAppEnvironmentsSearchResult>(url);