+++ OneBlink.SDK/models/FormElementLookup.cs
@@ -40,7 +40,10 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-
+        public long workspaceId
+        {
+            get; set;
+        }
     }
 
     public class FormElementLookupSearchResult : SearchResult