+++ OneBlink.SDK/models/FormsAppEnvironmentCloneOptions.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace OneBlink.SDK.Model
 {
     public class FormsAppEnvironmentCloneOptions